Added Basic Subsystem support and packet handling in both directions
The packet transmitter has been added, it currently supports asynchronous packet
transfers in both directions (client->server, server->client). Because of the new requirement
that the tpm blob is built on the server and signed on the client this is REALLY needed.

The subsystem infrastructure has also been added.
There is still a problem with the NamedPipeConnection in linux (i did not recognize that on windows)
where the NamedPipeListener is just says that a client connected but then is in an unusable state...
...very strange

Still missing is a sophisticated log framework. My research has resulted that
apache log4net (http://logging.apache.org/log4net/) is a very nice solution