Added UnixSocketConnection as replacement for named pipes on unix systems
As of mono 2.6 Named pipes are supported by the framework, but marked with a
todo, that they are not functional on unix systems.
The UnixSocketConnection implementation uses native unix socket, the drawback is
that now the project is not compileable on Microsoft .net framework because
it simply does not contain these unix socket implementations.
The simple solution would be to capsulate operating system specific stuff
into a external library which is loaded at runtime. The base library then only
contains platform independent code. Thats the way to go ;-)